pyproject.toml
[project]
name = "toy-ecom-site"
version = "0.1.0"
description = ""
authors = [
    {name = "Karishma Tank",email = "tankkarishma@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]
packages = [{include = "toy_ecom_site", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
